+++ packages/core-cairo/src/generate/erc721.ts
@@ -9,6 +9,7 @@ const booleans = [true, false];
 const blueprint = {
   name: ['MyToken'],
   symbol: ['MTK'],
+  baseUri: ['https://example.com/'],
   burnable: booleans,
   pausable: booleans,
   mintable: booleans,